+++ src/megan/clusteranalysis/indices/UniFrac.java
@@ -54,7 +54,7 @@ public class UniFrac {
             if (v.getOutDegree() != 1 && (Integer) v.getInfo() > 0)  // only use proper nodes
             {
                 countNodes++;
-                final int[] counts = (v.getOutDegree() == 0 ? viewer.getNodeData(v).getSummarized() : viewer.getNodeData(v).getAssigned());
+                final float[] counts = (v.getOutDegree() == 0 ? viewer.getNodeData(v).getSummarized() : viewer.getNodeData(v).getAssigned());
                 for (int s = 1; s <= distances.getNtax(); s++) {
                     for (int t = s + 1; t <= distances.getNtax(); t++) {
                         if ((counts[s - 1] < threshold) != (counts[t - 1] < threshold))
